src/agent_workflow.py
import argparse
import json
from google.cloud import bigquery
from google.cloud import language_v1
from google.cloud import aiplatform

# Simplified Agent Workflow
def agent_query(query, project_id):
    client_bq = bigquery.Client()
    query_job = client_bq.query("SELECT id, text FROM `your_dataset.your_table` LIMIT 10")
    docs = [dict(row) for row in query_job]

    if not docs:
        return {"response": "No documents found."}

    doc = docs[0]

    # Extract entities
    nlp_client = language_v1.LanguageServiceClient()
    document = language_v1.Document(content=doc['text'], type_=language_v1.Document.Type.PLAIN_TEXT)
    entities = nlp_client.analyze_entities(document=document)
    extracted = [(e.name, language_v1.Entity.Type(e.type_).name) for e in entities.entities]

    # Summarize
    aiplatform.init(project=project_id, location="us-central1")
    model = aiplatform.TextGenerationModel.from_pretrained("text-bison")
    summary = model.predict(doc['text'], max_output_tokens=128).text

    return {
        "query": query,
        "doc_id": doc['id'],
        "entities": extracted,
        "summary": summary
    }

if __name__ == "__main__":
    parser = argparse.ArgumentParser()
    parser.add_argument('--query', required=True)
    parser.add_argument('--project', required=True)
    args = parser.parse_args()

    response = agent_query(args.query, args.project)
    print(json.dumps(response, indent=2))
