requirements.txt
google-cloud-storage
google-cloud-bigquery
google-cloud-language
google-cloud-aiplatform
pandas
numpy
matplotlib
spacy
networkx
scikit-learn
tqdm
